Cleanup resource code and add unit tests (#61)
This is a general cleanup of all the resource code and related unit tests.

Resources are the basic building block of the scheduler. The resource objects are defined as  sparse objects in which all non existing quantities are considered zero. Quantities are limited to 64 bit integer values. Multiplication and addition of resources and their defined quantities is overflow protected. However overflows are only logged and return values are limited to either the minimum or maximum value.
All functions using resources are sparse resource and nil safe.