Linter integration (#80)
This change add linter checks to the code and formats the project using the standard for GO. Formatting fixes consist of: gofmt and goimports over all source files.
On top of that the most obvious lint issues raised by some of the linters have been fixed. Lint fixes consist of: simple, unused, error, static, shadowing and critic lint checkers. It also includes the standardisation of the capitalisation of names (ID etc). The use of non crypto rand was removed (gosec).
Make, configuration and IDE integration are provided in the documentation
All copyrights are update to 2020.