Recover existing allocations in the correct queue (#82)
During recovery, existing allocations should inherit the queue name from the application. When the application is recovered and placed via rules combined with recovered allocations that are updated by the admission controller. The queue in the allocation might differ from where the application runs. 
Allocations proposals should also not consider a node if there are not enough resources and thus fail fast.

Co-authored-by: Wilfred Spiegelenburg <wspiegelenburg@cloudera.com>