[YUNIKORN-14] add rest API for history (#109)
Add a reset API that provides historical information for containers and
applications running. It provides the count of the containers and
applications running for the last 24 hours at a 1 minute interval (not
configurable). If no data is available or no data is collected for the
specific counter a -1 is returned.

Fixes: #109