[YUNIKORN-99] Add state based app sorting (#145)
To allow FIFO scheduling to work when lots of applications are submitted
an extra state is added to the applications. This state shows a newly
starting application with just one allocation. If more allocations are
added the application goes into a running state. The starting state
times out after 5 minutes to prevent starvation of other apps.

A new sorting policy for applications in a queue is added which only
allows one starting application to be scheduled at a time. If there are
no starting applications in the list a new application without
allocations, accepted state, is added to the list of apps to schedule.

The sorting policy can be set on a leaf queue by specifying "statusaware"
as the sort policy.

Fixes: #145