[YUNIKORN-169] panic in ask release (#152)
* [YUNIKORN-169] panic in ask release

The panic appears when the allocation ask is removed before we can confirm the allocation in the scheduler. At that point, the cache has been updated with the allocation and the scheduler is finalizing the allocation in the scheduler.

The shim has been send and updates with the fact that the allocation has happened. This is part of the cache update at the same time as we update the scheduler. Before the shim has processed that update it still sees the ask as outstanding. It then cancels the ask, reason unknown, and not relevant for the issue. The scheduler processes ask cancellation in a separate goroutine compared to the main scheduler processing. The confirmation and the cancellation are now racing.
 - If the cancellation wins the confirmations panics.
 - If the confirmation wins there is no issue and the shim has been given an allocation which it can cancel.

The fix is to ensure the both allocations in the cache and the scheduler are released in this scenario. When the scheduler is unable to confirm an allocationProposal due to the request has been removed, it notifies the cache to release that allocation.