[YUNIKORN-199] Set unmanaged queue max resource
Set the unmanaged max queue resource based on the tag that is passed in
with a new application. The tag represents the quota from a namespace
from k8s. The max resource is only set on an unmanaged leaf queue as
the functionality should be used with the tag based placement rule.

This works around the hard namespace quota in k8s without the need for
major changes on the side of k8s namespace management.

The tag that is used is "namespace.resourcequota".