[YUNIKORN-222] Race in state change (#168)
If the application allocation is not confirmed before the ask is updated
the state is changed from accepted to running (via waiting) instead of
accepted to starting. This breaks the stateaware scheduling policy of a
queue.
Not affected by this issue is the removal of an ask, just the update of
the repeats is affected.

The ask repeat update state change check was racing with the allocation
confirmation from the cache. The ask repeat update could be triggered
before the confirmation. This only affects applications in the accepted
state. An application in accepted state was left with no pending ask,
no confirmed allocation and always an inflight allocation. This made
the triggering the state change ambiguous: there was no set of values
that always showed the correct state.

Fixes: #168