[YUNIKORN-241] add request sorting policy
Add a request sorting policy per queue for requests in an application.
The policy is set on a leaf queue using "request.sort.policy". The policies
supported are FIFO (default) and priority.
FIFO sorts based on the create time of the request. Priority sorts requests
in a descending order. Requests that have the same priority are sorted FIFO
within the set of requests with teh same priority.

Priorities range from 0 (lowest) to 2147483647 (maximum).

Internal change: application submit time is no longer stored as an integer
but as a time object equivalent to requests.