[YUNIKORN-251] reuse UUID for recovered allocations (#179)
When the shim recovers nodes it also recovers the allocations on the
nodes. Those allocations already get a UUID from the shim and follow a
different allocation path.
The UUID should be reused when it the allocation is recovered otherwise
the shim and core will be out sync. Recovered allocations are not
confirmed to the shim.
Allocations that are recovered must contain a UUID otherwise the
tracking on the core side will break.

Co-authored-by: Wilfred Spiegelenburg <wilfreds@apache.org>

Fixes: #179