[YUNIKORN-231] update pod condition for scale up
To support a cluster that can grow or shrink (auto scale) we need to
provide information back to an auto scaler. Auto scaling can not just be
triggered by an outstanding request as we have queue quotas to consider.

We should only mark a request to trigger an auto scale if the queue, and
the parent queues, can still accommodate the request within the max queue's
limits.
If enough requests have been marked to trigger an upscale changes to the
application or request ordering do not cause new requests to be marked for
upscaling as that can lead to over provisioning of nodes.