[YUNIKORN-231] update pod status based on quota (#173)
Support for auto-scaling with a single level queue hierarchy. Requirs the
use of queue limits and no limits on the K8s side.
Mark pending pods to trigger the scaling up as long as the queue’s quota is
sufficient. It limits the over provisioning of nodes, but does not
guarantee that there is no over provisioning.

Fixes: #173