[YUNIKORN-278] Support upscaling for unlimited queues (#184)
If a queue has no limits the headroom returned will be nil. This is not
the same as a zero resource. No resources will fit in a nil resource
which prevents upscaling any queue that has no limit set in the
hierarchy.
The fitin check will see all missing resources as zero (sparse resource
defintion)

Fixes: #184