[YUNIKORN-287] reserved ask release double release (#185)
When an ask gets released that is reserved and being allocated the
counters tracking reservations can be updated twice for one ask release.
The counters keep track of the number of reservations on a queue or
partition.

This can lead to other reservations being ignored. Nodes that have been
reserved by other asks for the same app will be skipped during
scheduling. In a small cluster this could lead to resource starvation

Fixes: #185