[YUNIKORN-273] expose config via rest
Expose the running config via a web call in both yaml (default) and json
(query parameter "json" required). The end point is "/ws/v1/config".

The yaml response is a multi line marshalled yaml object.
The json response is a one line marshalled json object.
The sha512 checksum is provided as part of the response. The checksum is
not part of the marshalled data. The checksum is printed as a 32 byte
hexadecimal sha256 string.