[YUNIKORN-273] expose config via rest (#194)
Expose the running config via a web call in both yaml (default) and json
(Accept header "application/json" required).
The end point is "/ws/v1/config" with method "GET"..

The yaml response is a multi line marshalled yaml object.
The json response is a one line marshalled json object.
The sha512 checksum is provided as part of the response. The checksum is
not part of the marshalled data. The checksum is printed as a 32 byte
hexadecimal sha256 string.