[YUNIKORN-390] Panic in placement rule evaluation (#201)
Placement rules can specify an arbitrary parent rule and a create flag for
the parent rule. If the parent rule is configured with the create flag set
to "true" and returns a non-existing queue the scheduler panics.

The fact that the parent queue does not exist means it cannot be a leaf
queue and the isLeafQueue() check should be skipped.
Adding tests to cover the parent rule returns to all rules.

Fixes: #201