[YUNIKORN-366] Add rest API to update queue configuration (#198)
This PR adds a Rest API to update the scheduler configuration, the endpoint is
POST "ws/v1/config". The scheduler validates the configs sent via the HTTP request
and if the configs are valid, it applies the configs and writes the content back to
the configmap; otherwise no change will be made and the request will be rejected.