[YUNIKORN-404] Handle updating mix case queue names (#206)
Queue names are not case sensitive.
When updating a queue with mixed cases in the config the queue was not
tagged correctly as still in the config. This would cause a queue with a
mixed case name to be removed during an update while it was present in
the configuration.

Fixes: #206