[YUNIKORN-480] root queue max resource update
When updating the configuration the root queue max resources should not
be set to the value specified in the config. The root queue represents
the total resources available on the cluster and is updated when nodes
are added removed.
Existing tests in the queue already cover this as part of the test queue
creation. A new test is added to the partition tests to cover the config
update.