[YUNIKORN-481] Deadlock when removing allocation ask (#228)
When removing an allocationAsk the partition was locked before the
application scheduling itself does it the other way around which can
cause a deadlock if the ask is removed while the application is
scheduled.
Possible deadlock in the node usage calculation removed.

Removal of an allocation contains a possible deadlock which has not been
handled as YUNIKORN-461 is open for changing that behaviour already.

E2E tests run as part of the k8shim cover the testing.

Fixes: #228