[YUNIKORN-476] placeholder allocation and swap (#236)
The core placeholder swapping logic implementation, the core looks at the placeholders and
actively replace them with the real pods from the app that has gang scheduling requirements.
The core tries to reuse the placeholder's space for the real pod first, if any of the placeholders
don't fit, it will try to find if any node in the cluster can fit the pod. If an alternative node is found,
it allocates the pod to that node and releases a placeholder.