[YUNIKORN-461] remove allocations list from partition
The allocations map in the partition is not needed anymore after the
cache removal. Updating and keeping the map in sync with the other
objects in the scheduler can cause a deadlock.

Remove the map and replace it with a simple counter for the web UI.
Simplify locking where possible to remove partition locks.