[YUNIKORN-484] Added waiting state timeout (#230)
An application goes into a waiting state when no allocations or asks are
running or waiting. This means the application has finished doing its
work. The application will now automatically transition to the
"Completed" state after 30sec.
Applications in the completed state are not linked to a queue anymore.
Dynamics queues can now also be cleaned up.

Fixes: #230