[YUNIKORN-548] move completed apps to separate map (#252)
Allow application ID reuse after the app is completed. When an application
state changes to completed applications are moved into a separate map. To
allow multiple this process to happen multiple times the key used to
store the application is changed. The application object itself is not
changed.
The new key is: <appID>-<finish time in unix timestamp format>

Fixes: #252