[YUNIKORN-557] log ambiguous sync message at DEBUG level (#254)
Allocation changes are communicated in two ways between the core and the
shim. When an allocation is released this duplicate communication can
cause a "sync failed" message to be logged. The order in which the
communication is processed might the failure not an error but expected
behaviour.
To prevent confusion we should not log this as an ERROR but as DEBUG
until we fix the communication.

This mechanism is being reviewed in YUNIKORN-462. This change needs to
be re-evaluated when that jira is fixed.

Fixes: #254