[YUNIKORN-564] StateAware node scale up improvement (#255)
The StateAware policy removes all applications but one for accepted and
starting states. This means that the upscaling checks do not look ahead
more than one application. Only a sub set of the resources is thus
counted and marked for scaling up, delaying the upscaling.
As scaling up nodes is slow and scheduling is fast the core should
prioritse scale by not excluding any applications from scale up checks
while using StateAware sorting for a queue.

Fixes: #255