[YUNIKORN-578] dynamic queue quota for gang size (#259)
A dynamic queue which is linked to a namespace and the quota linked to
that fails to properly enforce gang request size against that quota.
This only happens for the first application submitted to the namespace
and queue. The order for queue create, app addition and quota checks is
incorrect.

Fixes: #259