[YUNIKORN-574] Wait for placeholder cleanup (#260)
With this fix a new state is introduced: Failing state. Also, Waiting state is renamed to Completing.
When the scheduler cleans up the placeholders, the core notifies the shim to release allocations/asks wait
and wait for the confirmation before moving the application into the terminating state (Completed or Failed).