[YUNIKORN-586] enhance placeholder cleanup on timeout
When the placeholder processing times out we could have placeholders in
two states: released or not released.
In the clean up we need to account for that. Placeholders that are marked
for release have a real allocation linked and are being processed by the
shim. The timeout should not release those placeholders.
This also needs to happen when the application is in Completing state.
On confirmation by the shim these placeholder replacements will trigger
a state change back to Running for the application. The placeholders
should thus no longer be available.