[YUNIKORN-418]  Add config REST API (#231)
A new rest API to validate the scheduler configuration
Description: Validates the configuration. Returns true if it is a valid one, false and the reason if it is not allowed.
URL: /ws/v1/conf
Method: POST

Fixes: #231

In addition, response middleware has been added to return error also in json format and used httptest package for unit testing the same