[YUNIKORN-639] enhanced StateAware gang scheduling
In a queue with sort policy that is state aware there is a chance that
the state order is broken. There is a small delay between allocating
the last placeholder and getting the real allocation from the shim.
This allows a non gang app to be scheduled before the gang app. Mainly
causing issues if the non gang app only has one allocation and the
Starting state needs to time out.
Fixing this by moving a gang app to Starting when all placeholders are
allocated.