[YUNIKORN-405] Added checksum validation (#207)
Add checksum validation when the configuration is updated. The layout of
the config exposed in the REST call has also changed to accomodate this.
The checksum in the config is calculated over all data except for the
checksum itself.
The line starting with "checksum: " and the following 64 bytes, or until
the newline char are excluded from the checksum.
The checksum is calculated as a a SHA sum 256.

Fixes: #207