[YUNIKORN-193] Template for dynamic queues (#296)
A configured (managed) parent queue can now specify a template of
properties to be applied to a child queue that is dynamically created
below it in the hierarchy. A limited set of properties of the queue is
supported as part of the template.
The template is inherited as is from parent to parent if no new template
is specified for the parent.

This functionality superseedes and deprecates the functionality
implemented via YUNIKORN-195

Fixes: #296