[YUNIKORN-807] Improve performance of node sorting. (#307)
Add a node collection object that holds up all the nodes, it uses a
btree to store all the nodes and does the incremental updates when
necessary, the sorting is based on the node-score policy which is configurable. This can significantly improve the node-sorting performance.
Before this patch, the node-sorting is done per scheduling cycle, and
doing full-nodes-sorting every time.