[YUNIKORN-847] Use resource weighting for node sorting. (#323)
This is a rework of the node sorting algorithm, which brings some notable changes to the node sorting policies. This PR introduces "weights" for different resource types, default cpu and memory both have equal weight 1. The weight can be any positive fraction numbers, users can configure them differently according to the different requirements. The node score will be calculated based on the usage ratio per resource type in the white list (weight>0), and nodes will be sorted base on the score. This fixed the issue we saw in the past that does the calculation based on the largest number from one resource type.