[YUNIKORN-1093] Track rejected applications (#380)
Rejectde applications disappear from the core and cannot be tracked. The
only way to find the reason why is if an event is added to the pod. We
now track rejected applications in the core with the reason why itt was
rejected.

The list is exposed in the partition object as part of the REST call.
Rejected applications are retained iand cleaned up like completed
applications.

Closes: #380

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>