[YUNIKORN-1242] gang: delay in scheduling after using all placeholders
In the tryAllocate a check fails to take into account that an
application could request more real allocations than defined in the gang
request.
If multiple gang types are used this could cause a delay in scheduling
the "extra" requests for that gang type. Worst case scenario that could
be the placeholder timeout.