[YUNIKORN-1395] Account for preemptions in the placeholder data (#447)
If we preempt a placeholder to make room for a daemon set pod we do not
update the counter inside the application placeholder data object. This
causes the scheduler to think it needs to replace a placeholder but
never finds one to replace causing the real allocation to never
succeed

Same fix for placeholders that get removed as part of a node removal.

Closes: #447

Signed-off-by: Craig Condit <ccondit@apache.org>