[YUNIKORN-790] Implement MaxApplications for queues
Applications are considered running when they have allocations assigned.
The first state considered is the Starting state. Applications during
gang placeholder allocations stay in Accepted state but are counted as
part of the running count.
Max applications is configurable in the hierarchy. Configuration checks
are added and enforced. Tests are added for the configuration checks.

Logic for the tracking, updating and enforcement is located inside the
queue code. Unit tests are added for the logic.
Allocation cycle test is added as part of the partition tests.