[YUNIKORN-1445] partition cleanup: refactor code (#466)
The cleanup code needs a "cleanup"
Refactoring of cleanupExpiredApps to rely on just application ID lists.
Make GetAppsByState, GetRejectedAppsByState & GetCompletedAppsByState
internal to the partition and change the return value

Remove GetAppsInTerminatedState and GetTotalCompletedApplicationCount as
they are unused or only for test and not needed.
Cleanup tests to use assert.Equal instead of asser.Assert

Closes: #466

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>