[YUNIKORN-1590] REST interface shows incorrect placeholder data (#509)
The REST interface does not initialise all fields for the placeholders
correctly. The allocation DAO did not set the Placeholder or
TaskGroupName. The Ask did not expose the originator flag.

Allocation code was duplicated for nodes and applications. This could
lead to inconsistent data to be shown for the same object in one dump.

Closes: #509

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>