[YUNIKORN-1658] autoscaling support with unlimited resources (#523)
* internalHeadRoom of the queue calculates incorrectly if allocations
have an unlimited resource type as the unlimited resource turns negative
in the queue headroom.
* getOutstandingRequests from the application breaks if a request has a
resource type that is not limited in the queue hierarchy. This will
cause the "fit in" check to fail as the unlimited resource turns
negative in the headroom.

Both issues result in a scale up failure.

Closes: #523

Signed-off-by: Craig Condit <ccondit@apache.org>