[YUNIKORN-1916] Align headroom and resource tracking resolution logic (#618)
Headroom and increase need to run the exact same logic for user and
group tracker create and checks. The wildcard group is a separate group
tracker that gets returned based. Wildcard groups are only tracked in a
leaf and all use the same group tracker object.

Only calculate headroom once for each application in the scheduling
cycle not for each ask.

Closes: #618

Signed-off-by: Manikandan R <manirajv06@gmail.com>