[YUNIKORN-1920] headroom with parent queue usage
Headroom needs to take into account the usage in each queue and not
assume that the full max resource is available if the queue is not the
leaf queue.
Add a queue traker object for any queue that is not found while
traversing the hierarchy. The user/group has an application running in
that queue and will add usage at some point.

Add tests to cover usage and max with different resource types set.