[YUNIKORN-1993] race condition in allocation removal
Race between allocation removal and the timed transition to Completed
state change in the partition code.
If the removal of an allocation is in progress when the state transition
causes the queue removal (entry of the Completed state) the queue will
not be updated and the allocation will leak.

No test case as the timing required is not reproducible in unit or e2e
tests.