[YUNIKORN-1987] Identity config fields not set explicitly
Added functionality to identify configuration fields not set explicitly. This change will highlight the difference between fields that aren't set by the user versus those explicitly set to default value. New logic defines a `DefaultValues` struct and uses a switch case in the `CheckAndSetDefault` method to discern if the value is default (uint64 is max value, string is "defaultStringValue"). Configuration-specific UnmarshalYAML methods are utilized to facilitate the usage of aforementioned `CheckAndSetDefault` method. Added corresponding tests for the same.