[YUNIKORN-2025] Mismatched running container count (#675)
Preempted containers are counted twice as released. This causes the
running container count to be incorrect when preempting containers.

Handle PREEMPTED_BY_SCHEDULER just like TIMEOUT: do not send the release
back to the shim in removeAllocation. The removal was triggered in the
by the core in the first place.

Closes: #675

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>