[YUNIKORN-2087] reinstate node utilisation rest endpoint (#691)
The node utilisation endpoint is required to be exposed for the
dashboard to work after YUNIKORN-325
Moving the endpoint to /ws/v1/scheduler/node-utilization
The endpoint will need to change as part of YUNIKORN-2088

Dominant resource is calculated for the root queue and that utilisation
is shown for the nodes. The dominant reource type is the type with the
highest usage ratio. A pre defined ratio for zero usage and capacity is
defined as: 0 capacity, 0 usage: 0; 0 capacity, >0 usage: 1 otherwise
ratio = usage / capacity.

Closes: #691

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>