[YUNIKORN-2171] race between node removal and scheduling (#724)
When a node gets removed the partition resources and thus the root max
resources are decreased. Cleanup of the allocations happens after that.
This means that for a short period of time the root queue max resources
are already decreased while the usage is not.

If the queue headroom is limited by the root queue then we could have a
race between the removal of the node allocations and scheduling adding a
new allocation. Scheduling could fail with the root queue usage being
over the max resources.

Closes: #724

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>