[YUNIKORN-2204] Use ask unique id for allocation (#740)
Do not generate a new UUID for an allocation when it gets added. Re-use
the ask key which is the same as the K8s pod UID to improve on the
troubleshooting experience.
The new allocationID will be the ask key followed by the current repeat
counter, seperated by a dash '-'

The UUID is renamed to allocationID on the allocation. Scheduler
interface changes will follow. Replaced GetUUID with GetAllocationID in
already modified lines

Closes: #740

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>