[YUNIKORN-2212] Outstanding request collection optimisation (#745)
Auto scaling n K8s relies on the pod to be marked unschedulable. That is
done via a regular callback from the core to the shim. The logic for
building the pod list in this callback is simple. This could cause over
scaling nodes or excessive K8s API server calls.

Improvements added:
* Don't collect requests that have not been scheduled yet.
* If already triggered scale up do not sent again
* Check for gang placeholders for any real ask
* Don't scale if pod has required node set (daemon set pod)

Track the state of scheduling and auto scale triggering.

Closes: #745

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>