[YUNIKORN-2406] Ensure shim is notified of existing allocations being added
When existing allocations are registered from the shim to the core, they are
added to the partition, but the shim is never notified that the allocations
were successfully registered. This manifests in the core as missing history
for container allocations, and in the shim by tasks not transitioning to a
bound state.