[YUNIKORN-2646] Allow lock order detection to be turned off
Most of the false positive deadlock detections are linked to the
preemption cycle. In that cycle we have a lock on the application being
scheduled and acquire read locks on other applications to look for
victims.
In the next scheduling cycle the applications involved might be reversed
triggering an inconsistent lock detection to be logged. That causes log
spew.
The lock order detection can be turned off by setting an option exposed
via a new environment variable.