[YUNIKORN-2789] Queue internalGetMax should use permissive calculator
In the implementation of internalGetMax on the queue, we call resources
ComponentWiseMin(). This returns 0 values for each type that is not
defined in both of the resources passed in.
In case there is no overlap it will return a resource with all types set
to "zero".
That is contrary to the documentation on how quotas work and what we do
everywhere besides the internalGetMax implementation.

The ComponentWiseMin function is only used in queue internalGetMax.
Refactor to make the ComponentWiseMinPermissive version the standard
(rename and remove "permissive"), and remove the existing
ComponentWiseMin code.