[YUNIKORN-2796] Prune zero resource type from root and partition
After the last node is removed that registers a specific resource type
the root queue and partition show it as a resource type with 0 maximum.
A resource type that does not exist should not be shown at all.
Pruning the 0 valued resource types for root and partition will fix that

Clean up of the partition code to use the same function to update
partition and root settings for create, update and delete of a node.

Fix root queue check for max resource.
Fix tests that fake allocation or node registration to correctly set the
root queue limits.