[YUNIKORN-2965] Move statedump and stack REST to debug endpoint
The REST calls to get a statedump and the stacks of the system are
exposed on the wrong endpoint. They belong in /debug and not in /ws/v1.
Both are for troubleshooting only and the content is not stable.

Placing a 301 redirect on the old endpoints for clients to follow.
The proxy build into the web UI does not proxy the debug endpoints.