[YUNIKORN-2967] Cleanup REST response headers (#994)
Only respond with the allowed methods for the request, not with a
general all allowed set. OPTIONS is supported via the generic config.
Add a test to make sure a change in router does not break that.

Remove the Access-Control-Allow-Credentials as recommended in the RFC.
We also do not use cookies or authentication so not relevant to set.

Closes: #994

Signed-off-by: Craig Condit <ccondit@apache.org>