[YUNIKORN-2978] Fix handling of reserved allocations where node differs
YUNIKORN-2700 introduced a bug where allocations of previously-reserved
tasks were not handled correctly in the case where we schedule on a
different node than the reservation. Ensure that we unreserve and
allocate using the proper node in both cases.

Also introduce additional logging of allocations on nodes to make
finding issues like this easier in the future.