[YUNIKORN-2925] Remove internal object from application REST info (#999)
Remove the TrackedResource object from the REST response for the
application. Moved the tracked resources into a separate object inside
the app DAO, as we have done for other complex pieces.
Cleanup exposure of the summary object to be testing only.
Add DAO retrieval of TrackedResource type for the web handler.

Closes: #999

Signed-off-by: Craig Condit <ccondit@apache.org>