[YUNIKORN-2974] Expose preemption and priority settings via REST (#995)
Add preemption and priority details to the queue REST information.
Currently the configuration is exposed via the properties. The
properties contain the raw configuration values which need to be
interpreted.

Invalid configuration values are ignored and not set on the queue. This
could cause an incorrect view of what is or should be active on the
queue if based on the properties. Using the active values from the
object to show the real state.

Expose Partition level preemption flag as part of this change to
complement the queue details.

Closes: #995

Signed-off-by: Craig Condit <ccondit@apache.org>